refactor: 이미지 폼 분리
