+++ src/pages/Articles/components/FormImage/FormImage.module.scss
@@ -1,0 +1,127 @@
+@use "src/utils/scss/media" as media;
+
+.images {
+  gap: 8px;
+  align-items: center;
+  box-sizing: border-box;
+
+  @include media.media-breakpoint(mobile) {
+    width: 100%;
+    height: auto;
+    padding: 0;
+  }
+
+  &__header {
+    display: flex;
+    flex-direction: column;
+  }
+
+  &__text {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  &__counter {
+    color: #727272;
+    font-family: Pretendard, sans-serif;
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 1.6;
+    text-align: center;
+  }
+
+  &__list {
+    width: 327px;
+    height: 128px;
+    box-sizing: border-box;
+    border-radius: 8px;
+    background-color: #f5f5f5;
+    display: flex;
+    gap: 8px;
+    padding: 12px 0 17px 12px;
+    margin: 16px 0 24px;
+    overflow-x: scroll;
+    overflow-y: hidden;
+    scrollbar-width: thin;
+    scrollbar-color: #999 transparent;
+
+    &::-webkit-scrollbar-thumb {
+      background: #d9d9d9;
+      border-radius: 4px;
+    }
+
+    @include media.media-breakpoint(mobile) {
+      max-width: calc(100vw - 48px);
+      width: 100%;
+    }
+  }
+
+  &__item {
+    position: relative;
+  }
+
+  &__image {
+    width: 112px;
+    height: 99px;
+    object-fit: cover;
+  }
+
+  &__delete {
+    position: absolute;
+    right: -8px;
+    top: -8px;
+  }
+
+  &__upload {
+    width: 327px;
+    height: 40px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 8px;
+    border-radius: 8px;
+    background-color: #e4f2ff;
+    color: #10477a;
+    font-family: Pretendard, sans-serif;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 1.6;
+
+    @include media.media-breakpoint(mobile) {
+      width: 100%;
+    }
+
+    &:hover {
+      cursor: pointer;
+    }
+
+    input {
+      display: none;
+    }
+  }
+}
+
+.title {
+  font-family: Pretendard, sans-serif;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 1.2;
+
+  @include media.media-breakpoint(mobile) {
+    font-size: 14px;
+    line-height: 1.6;
+  }
+
+  &__description {
+    color: #727272;
+    font-family: Pretendard, sans-serif;
+    font-size: 14px;
+    line-height: 1.6;
+
+    @include media.media-breakpoint(mobile) {
+      margin-bottom: 12px;
+      font-size: 12px;
+      line-height: 1.6;
+    }
+  }
+}
